Aggiunti pagine_min, pagine_max, valutazione_min e limitatore di risultati. Cambiata laggermente la UI.

+++ src/main/java/org/libri/esplora/backend/data/ControllerPrincipale.java
@@ -49,11 +49,11 @@ public class ControllerPrincipale {
     @GetMapping(path = "ricerca")
     public ResponseEntity<List<RisultatoRicerca>> ricerca(
             @RequestParam(defaultValue = "") String valore_ricerca,
-            @RequestParam(defaultValue = "-999999999") Year anno_min,  @RequestParam(defaultValue = "999999999") Year anno_max,
-            @RequestParam(defaultValue = "-999999999") Float prezzo_min,        @RequestParam(defaultValue = "999999999") Float prezzo_max,
-            @RequestParam(defaultValue = "1") Short pagine_min,        @RequestParam(defaultValue = "32767") Short pagine_max,
-            @RequestParam(defaultValue = "-1") Byte valutazione_min,
-            @RequestParam(defaultValue = "") String genere,            @RequestParam(defaultValue = "") String lingua) {
+            @RequestParam(defaultValue = "-999999999") Year anno_min,       @RequestParam(defaultValue = "999999999") Year anno_max,
+            @RequestParam(defaultValue = "-999999999") Float prezzo_min,   @RequestParam(defaultValue = "999999999") Float prezzo_max,
+            @RequestParam(defaultValue = "1") Short pagine_min,             @RequestParam(defaultValue = "32767") Short pagine_max,
+            @RequestParam(defaultValue = "-1") Double valutazione_min,
+            @RequestParam(defaultValue = "") String genere,                 @RequestParam(defaultValue = "") String lingua) {
         // TODO Controlli sui valori
 
         List<RisultatoRicerca> risultato = servizioRicerca.ricerca(valore_ricerca, anno_min, anno_max, prezzo_min, prezzo_max, pagine_min, pagine_max, valutazione_min, genere, lingua);
