+++ src/main/java/org/libri/esplora/backend/data/service/ServizioRicerca.java
@@ -17,7 +17,7 @@ public class ServizioRicerca {
         this.repository = repository;
     }
 
-    public List<RisultatoRicerca> ricerca(String valoreRicerca, Year annoMin, Year annoMax, Float prezzoMin, Float prezzoMax, Short pagineMin, Short pagineMax, Byte valutazioneMin, String genere, String lingua) {
+    public List<RisultatoRicerca> ricerca(String valoreRicerca, Year annoMin, Year annoMax, Float prezzoMin, Float prezzoMax, Short pagineMin, Short pagineMax, Double valutazioneMin, String genere, String lingua) {
         return repository.ricerca(valoreRicerca, annoMin, annoMax, prezzoMin, prezzoMax, pagineMin, pagineMax, valutazioneMin, genere, lingua);
     }
